requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
reportlab==4.0.7